fix RIDER-123082 Implement automatic selection of Editor DLL matching Unity's version
A utility function has been added in UnityVersionUtils.cs that enables the automatic selection of the Editor DLL that most closely matches Unity's current version. This functionality has been integrated into RiderInitializer.cs, and automatically selects the DLL file with the closest match to Unity's version, providing a better seamless integration experience. On failure to find a suitable version, it reverts to the original strategy of going with the default EditorPluginAssemblyName .dll.